Refactor project to rely on categories for deserialization.
This both produces a better code layout, and vastly increases extensibility.
This also means we can support deserializing into existing instances.

The mapperconfig preProcBlock was removed along the way. I'm no longer
convinced it was the level for the solution to exist. This is better
handled by whatever is handing over the serialized data.